+++ gridopt/power_flow/augl_pf.py
@@ -1,17 +1,15 @@
 #*****************************************************#
 # This file is part of GRIDOPT.                       #
 #                                                     #
-# Copyright (c) 2015-2016, Tomas Tinoco De Rubira.    #
+# Copyright (c) 2015-2017, Tomas Tinoco De Rubira.    #
 #                                                     #
 # GRIDOPT is released under the BSD 2-clause license. #
 #*****************************************************#
 
 from __future__ import print_function
-import pfnet
 import numpy as np
 from .method_error import *
 from .method import PFmethod
-from optalg.opt_solver import OptSolverError,OptCallback,OptTermination,OptSolverAugL
 
 class AugLPF(PFmethod):
     """
@@ -19,32 +17,40 @@ class AugLPF(PFmethod):
     """
 
     name = 'AugLPF'
-
-    parameters = {'weight_vmag':1e0,     # for reg voltage magnitude penalty
-                  'weight_vang':1e-3,    # for angle difference penalty
-                  'weight_pq':1e-3,      # for gen powers
-                  'weight_t':1e1,        # for tap ratios
-                  'weight_b':1e-4,       # for shunt susceptances
-                  'lock_taps':True,      # flag for locking transformer tap ratios
-                  'lock_shunts':True,    # flag for locking swtiched shunts
-                  'vmin_thresh':0.1,     # threshold for vmin
-                  'limit_gens':True}     # flag for limiting generator reactive power (PV-PQ switching)
-
+        
+    parameters = {'weight_vmag': 1e0,  # for reg voltage magnitude penalty
+                  'weight_vang': 1e0,  # for angle difference penalty
+                  'weight_pq': 1e-3,   # for gen powers
+                  'weight_t': 1e-3,    # for tap ratios
+                  'weight_b': 1e-3,    # for shunt susceptances
+                  'lock_taps': True,   # flag for locking transformer tap ratios
+                  'lock_shunts': True, # flag for locking swtiched shunts
+                  'limit_gens':True,   # flag for limiting generator reactive power (PV-PQ switching)
+                  'feastol' : 1e-4,    # see AugL
+                  'optol' : 1e-4,      # see AugL
+                  'kappa' : 1e-5,      # see AugL
+                  'vmin_thresh': 0.1}  # threshold for vmin
+                  
     def __init__(self):
 
+        from optalg.opt_solver import OptSolverAugL
+
         PFmethod.__init__(self)
-        self.parameters = AugLPF.parameters.copy()
-        self.parameters.update(OptSolverAugL.parameters)
+        parameters = OptSolverAugL.parameters.copy()
+        parameters.update(AugLPF.parameters)
+        self.parameters = parameters
 
     def create_problem(self,net):
 
+        import pfnet
+
         # Parameters
         params = self.parameters
-        weight_vang = params['weight_vang']
-        weight_vmag = params['weight_vmag']
-        weight_pq = params['weight_pq']
-        weight_t = params['weight_t']
-        weight_b = params['weight_b']
+        wm = params['weight_vang']
+        wa = params['weight_vmag']
+        wp = params['weight_pq']
+        wt = params['weight_t']
+        wb = params['weight_b']
         lock_taps = params['lock_taps']
         lock_shunts = params['lock_shunts']
         limit_gens = params['limit_gens']
@@ -60,10 +66,6 @@ class AugLPF(PFmethod):
                       'variable',
                       'not slack',
                       ['voltage magnitude','voltage angle'])
-        net.set_flags('bus',
-                      'variable',
-                      ['not slack','regulated by generator'],
-                      'voltage magnitude deviation')
         net.set_flags('generator',
                       'variable',
                       'slack',
@@ -75,64 +77,47 @@ class AugLPF(PFmethod):
 
         # Tap ratios
         if not lock_taps:
-            net.set_flags('bus',
-                          'variable',
-                          'regulated by transformer',
-                          'voltage magnitude violation')
-            net.set_flags('branch',
+            net.set_flags('branch', 
                           'variable',
                           'tap changer - v',
-                          ['tap ratio','tap ratio deviation'])
+                          'tap ratio')
 
         # Shunt voltage control
         if not lock_shunts:
-            net.set_flags('bus',
-                          'variable',
-                          'regulated by shunt',
-                          'voltage magnitude violation')
-            net.set_flags('shunt',
+            net.set_flags('shunt', 
                           'variable',
                           'switching - v',
-                          ['susceptance','susceptance deviation'])
+                          'susceptance')
 
         try:
             num_vars = (2*(net.num_buses-net.get_num_slack_buses()) +
-                        2*(net.get_num_buses_reg_by_gen()-net.get_num_slack_buses()) +
                         net.get_num_slack_gens() +
                         net.get_num_reg_gens())*net.num_periods
             if not lock_taps:
-                num_vars += (2*net.get_num_buses_reg_by_tran() +
-                             3*net.get_num_tap_changers_v())*net.num_periods
+                num_vars += net.get_num_tap_changers_v()*net.num_periods
             if not lock_shunts:
-                num_vars += (2*net.get_num_buses_reg_by_shunt() +
-                             3*net.get_num_switched_shunts())*net.num_periods
-            if not lock_shunts and not lock_taps:
-                num_vars -= 2*len([b for b in net.buses
-                                   if (b.is_regulated_by_tran() and
-                                       b.is_regulated_by_shunt())])*net.num_periods
+
+                num_vars += net.get_num_switched_shunts()*net.num_periods
             assert(net.num_vars == num_vars)
         except AssertionError:
-            raise PFmethodError_BadProblem(self)
-
+            raise PFmethodError_BadProblem(self)  
+            
         # Set up problem
-        problem = pfnet.Problem()
-        problem.set_network(net)
-        problem.add_constraint('AC power balance')
-        problem.add_constraint('generator active power participation')
-        problem.add_constraint('generator reactive power participation')
+        problem = pfnet.Problem(net)
+        problem.add_constraint(pfnet.Constraint('AC power balance',net))
+        problem.add_constraint(pfnet.Constraint('generator active power participation',net))
+        problem.add_constraint(pfnet.Constraint('generator reactive power participation',net))
         if limit_gens:
-            problem.add_constraint('voltage regulation by generators')
+            problem.add_constraint(pfnet.Constraint('voltage regulation by generators',net))
+            problem.add_function(pfnet.Function('generator powers regularization',wp/max([net.num_generators,1.]),net))
+        problem.add_function(pfnet.Function('voltage magnitude regularization',wm/max([net.num_buses,1.]),net))
+        problem.add_function(pfnet.Function('voltage angle regularization',wa/max([net.num_buses,1.]),net))
         if not lock_taps:
-            problem.add_constraint('voltage regulation by transformers')
+            problem.add_constraint(pfnet.Constraint('voltage regulation by transformers',net))
+            problem.add_function(pfnet.Function('tap ratio regularization',wt/max([net.get_num_tap_changers_v(),1.]),net))
         if not lock_shunts:
-            problem.add_constraint('voltage regulation by shunts')
-        problem.add_function('voltage magnitude regularization',weight_vmag)
-        problem.add_function('voltage angle regularization',weight_vang)
-        problem.add_function('generator powers regularization',weight_pq)
-        if not lock_taps:
-            problem.add_function('tap ratio regularization',weight_t)
-        if not lock_shunts:
-            problem.add_function('susceptance regularization',weight_b)
+            problem.add_constraint(pfnet.Constraint('voltage regulation by shunts',net))
+            problem.add_function(pfnet.Function('susceptance regularization',wb/max([net.get_num_switched_shunts(),1.]),net))
         problem.analyze()
 
         # Return
@@ -141,7 +126,7 @@ class AugLPF(PFmethod):
     def get_info_printer(self):
 
         def info_printer(solver,header):
-            net = solver.problem.network
+            net = solver.problem.wrapped_problem.network
             if header:
                 print('{0:^5}'.format('vmax'), end=' ')
                 print('{0:^5}'.format('vmin'), end=' ')
@@ -160,6 +145,8 @@ class AugLPF(PFmethod):
 
     def solve(self,net):
 
+        from optalg.opt_solver import OptSolverError, OptCallback, OptTermination, OptSolverAugL
+        
         # Parameters
         params = self.parameters
         vmin_thresh = params['vmin_thresh']
@@ -171,7 +158,7 @@ class AugLPF(PFmethod):
 
         # Termination
         def t1(s):
-            if np.min(s.problem.network.bus_v_min) < vmin_thresh:
+            if np.min(s.problem.wrapped_problem.network.bus_v_min) < vmin_thresh:
                 return True
             else:
                 return False
@@ -200,32 +187,35 @@ class AugLPF(PFmethod):
             self.set_dual_variables(solver.get_dual_variables())
             self.set_net_properties(net.get_properties())
             self.set_problem(problem)
-
+           
+            # Restore net properties
+            net.update_properties()
+ 
     def update_network(self,net):
 
         # Get data
         problem = self.results['problem']
-        x = self.results['primal_variables']
-        lam,nu,mu,pi = self.results['dual_variables']
-
+        x = self.results['primal variables']
+        lam,nu,mu,pi = self.results['dual variables']
+       
         # No problem
         if problem is None:
             raise PFmethodError_NoProblem(self)
 
         # Checks
         assert(problem.x.shape == x.shape)
-        assert(net.num_vars == x.size)
+        assert(net.num_vars+problem.num_extra_vars == x.size)
         assert(problem.A.shape[0] == lam.size)
         assert(problem.f.shape[0] == nu.size)
-        assert(mu is None or not mu.size)
-        assert(pi is None or not pi.size)
+        assert(problem.G.shape[0] == mu.size)
+        assert(problem.G.shape[0] == pi.size)
 
         # Network quantities
-        net.set_var_values(x)
+        net.set_var_values(x[:net.num_vars])
 
         # Network properties
         net.update_properties()
 
         # Network sensitivities
         net.clear_sensitivities()
-        problem.store_sensitivities(lam,nu,mu,pi)
+        problem.store_sensitivities(lam,nu,np.zeros(0),np.zeros(0))